reorganize app folders
