+++ app/db/sql/user/delete.sql
