+++ app/services/UserDAO.js
@@ -1,5 +1,5 @@
-const db = require('../config/db');
-const sql = require('../config/sqlProvider').users;
+const db = require('../db/db');
+const sql = require('../db/sqlProvider').users;
 const User = require('../models/User');
 
 class UserDAO {
