+++ server.js
@@ -1,16 +1,20 @@
-if (!process.env) {
-  require('dotenv').config();
-}
-
-process.env.ENV = process.env.ENV || 'dev';
 
 const path = require('path');
 const express = require('express');
 const webpack = require('webpack');
 const webpackDevMiddleware = require('webpack-dev-middleware');
 const webpackHotMiddleware = require('webpack-hot-middleware');
+
 const config = require('./webpack.config');
 const app = require('./app/app');
+
+if (!process.env) {
+  require('dotenv').config();
+}
+
+process.env.ENV = process.env.ENV || 'dev';
+
+const port = process.env.PORT;
 
 if (process.env.ENV === 'dev') {
   const compiler = webpack(config);
@@ -25,12 +29,12 @@ if (process.env.ENV === 'dev') {
   app.use(webpackHotMiddleware(compiler));
 }
 
-app.use(express.static(path.join(__dirname, '/dist')));
+app.use(express.static(path.join(__dirname, '/public')));
+
 app.get('/', (request, response) => {
   response.sendFile(path.join(__dirname, '/dist/index.html'));
 });
 
-const port = process.env.PORT;
 app.listen(port, () => {
   console.log(`${process.env.ENV.toUpperCase()}: Listening on port ${port}`);
 });
